Multi-project tests + remove "impliesSubProjects"
- Add multi-project tests with sample projects.
- Remove impliesSubProjects=true from taskTree definition.
When set to true, task tree is shown for the current project
only, and not for sub-projects.
Users can add this themselves in build.gradle if they wish.
While this might be a better default, earlier versions of
the taskTree plugin did not have impliesSubProjects=true.
So I rather avoid breaking existing behviour.
- Remove oraclejdk7 from travis configuration.
tests running on oraclejdk7 on Travis CI encounter
OutOfMemoryError: Permgen space